requirements.txt
alembic==0.9.7
click==6.7
Flask==0.12.2
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
psycopg2==2.7.4
python-dateutil==2.6.1
python-decouple==3.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.2
Werkzeug==0.14.1
